// src/app/model/ains-details.ts
export class AinsDetails {
    animal_Tag_Number:string;
    bull_Name: string;
    date_Of_Event: Date;
    semen_Type: string;
    bull_Number: string;
    ai_Center_Id: number;
    technique_Name: string;
    bull_Breed_Type_Id: number;
    technician_Name: string;
    ai_Technician_Type: string;
    type_Of_Assited_Formula: string;
    date_Of_Pd_Status: Date;
    pd_Status_Date: Date;
    breeding_Status: string;
    longitude: number;
    user_Gewog_Id: number;
    type_Of_Event: string;
    pd_Status: string;
    bull_Exotic: number;
    latitude: number;
    altitude:number;
    pd_Date: Date;
    remarks: string;
}